breakout slack intergration into and encapsulated module
